Prepare for creating script to convert radar markdown to JSON
- Removed "module" type from package.json as it is not recommended by
  https://www.typescriptlang.org/docs/handbook/modules/guides/choosing-compiler-options.html#im-using-a-bundler
- Split outconfig types from app & radar config itself
- Split radar/io into mdio & jsonio.
  - mdio is solely used during the build process
  - jsonio will be used by the client component to fetch the JSON
    data created during the build step
- Following Cypress guidelines and creating a separate tsconfig
  to avoid clashing types with Jest. tsc --noEmit is now clean
  but there is a slight issue that VSCode doesn't know which
  tsconfig to use.
